-Add crashlytics -Fix numerous bugs -Add Stripe live key w/compiler macros -Fix 3.5 inch screen issues

+++ RyteBytes/OrderItem.h
@@ -20,6 +20,6 @@
 -(id)initWithMenuItem:(MenuItem*)item;
 -(id)initWithMenuItem:(MenuItem*)item withQuantity:(int)q;
 
--(double)calculateCost;
+-(float)calculateCost;
 
 @end
